Telegram-Scrapper-main/requirements.txt
git+https://github.com/Altruix/pyrogram
python-dotenv
pyromod
cryptg
aiofiles